script to toggle deletion of tables before start

+++ src/server.js
@@ -17,16 +17,23 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 const db = require('./models');
-db.sequelize.sync({ force: true }).then(() => {
-    console.log('Drop and re-sync db');
-});
-// db.sequelize.sync()
-//     .then(() => {
-//         console.log('Synced DB');
-//     })
-//     .catch(err => {
-//         console.error(err);
-//     });
+
+var myArgs = process.argv.slice(2);
+if (myArgs[0] == 'c') {
+
+    db.sequelize.sync({ force: true }).then(() => {
+        console.log('Drop and re-sync db');
+    });
+}
+else {
+    db.sequelize.sync()
+        .then(() => {
+            console.log('Synced DB');
+        })
+        .catch(err => {
+            console.error(err);
+        });
+}
 
 
 // routes
